Move `ThriftMetadataWriter` and `ParquetMetadataWriter` to a new module
